Weapon and ContinuousWeapon
