+++ core/src/mindustry/type/weapons/RepairBeamWeapon.java
@@ -20,11 +20,7 @@ import mindustry.world.meta.*;
 
 import static mindustry.Vars.*;
 
-/**
- * Note that this weapon requires a bullet with a positive maxRange.
- * Rotation must be set to true. Fixed repair points are not supported.
- * */
-public class RepairBeamWeapon extends Weapon{
+public class RepairBeamWeapon extends TargetWeapon{
     public boolean targetBuildings = false, targetUnits = true;
 
     public float repairSpeed = 0.3f;
@@ -41,6 +37,7 @@ public class RepairBeamWeapon extends Weapon{
     //only for blocks
     public Color healColor = Pal.heal;
     public Effect healEffect = Fx.healBlockFull;
+    public float healEffectInterval = 1f;
 
     public RepairBeamWeapon(String name){
         super(name);
@@ -50,17 +47,11 @@ public class RepairBeamWeapon extends Weapon{
     }
 
     {
-        //must be >0 to prevent various bugs
-        reload = 1f;
-        predictTarget = false;
         autoTarget = true;
         controllable = false;
         rotate = true;
-        useAmmo = false;
-        mountType = HealBeamMount::new;
+        mountType = HealBeamMountR::new;
         recoil = 0f;
-        noAttack = true;
-        useAttackRange = false;
     }
 
     @Override
@@ -70,47 +61,30 @@ public class RepairBeamWeapon extends Weapon{
     }
 
     @Override
-    public float dps(){
-        return 0f;
+    public void drawWeapon(Unit unit, BaseWeaponMount mount, float wx, float wy, float weaponRotation){
+        super.drawWeapon(unit, mount, wx, wy, weaponRotation);
+
+        HealBeamMountR heal = (HealBeamMountR)mount;
+
+        if(unit.canShoot()){
+            float z = Draw.z();
+            RepairTurret.drawBeam(wx, wy, unit.rotation + mount.rotation, shootY, unit.id, heal.target == null || controllable ? null : (Sized)heal.target, unit.team, heal.strength,
+                pulseStroke, pulseRadius, beamWidth + Mathf.absin(widthSinScl, widthSinMag), heal.lastEnd, heal.offset, laserColor, laserTopColor,
+                laser, laserEnd, laserTop, laserTopEnd);
+            Draw.z(z);
+        }
     }
 
     @Override
-    public void load(){
-        super.load();
-
-        laser = Core.atlas.find("laser-white");
-        laserEnd = Core.atlas.find("laser-white-end");
-        laserTop = Core.atlas.find("laser-top");
-        laserTopEnd = Core.atlas.find("laser-top-end");
-    }
-
-    @Override
-    protected Teamc findTarget(Unit unit, float x, float y, float range, boolean air, boolean ground){
-        var out = targetUnits ? Units.closest(unit.team, x, y, range, u -> u != unit && u.damaged()) :  null;
-        if(out != null || !targetBuildings) return out;
-        return Units.findAllyTile(unit.team, x, y, range, Building::damaged);
-    }
-
-    @Override
-    protected boolean checkTarget(Unit unit, Teamc target, float x, float y, float range){
-        return !(target.within(unit, range + unit.hitSize/2f) && target.team() == unit.team && target instanceof Healthc u && u.damaged() && u.isValid());
-    }
-
-    @Override
-    protected void shoot(Unit unit, WeaponMount mount, float shootX, float shootY, float rotation){
-        //does nothing, shooting is handled in update()
-    }
-
-    @Override
-    public void update(Unit unit, WeaponMount mount){
+    public void update(Unit unit, BaseWeaponMount mount){
         super.update(unit, mount);
 
         float
-        weaponRotation = unit.rotation - 90,
-        wx = unit.x + Angles.trnsx(weaponRotation, x, y),
-        wy = unit.y + Angles.trnsy(weaponRotation, x, y);
+            weaponRotation = unit.rotation - 90,
+            wx = unit.x + Angles.trnsx(weaponRotation, x, y),
+            wy = unit.y + Angles.trnsy(weaponRotation, x, y);
 
-        HealBeamMount heal = (HealBeamMount)mount;
+        HealBeamMountR heal = (HealBeamMountR)mount;
         boolean canShoot = mount.shoot;
 
         if(!autoTarget){
@@ -143,50 +117,47 @@ public class RepairBeamWeapon extends Weapon{
             }
         }
 
-        heal.strength = Mathf.lerpDelta(heal.strength, Mathf.num(autoTarget ? mount.target != null : canShoot), 0.2f);
+        heal.strength = Mathf.lerpDelta(heal.strength, Mathf.num(autoTarget ? heal.target != null : canShoot), 0.2f);
 
         //create heal effect periodically
-        if(canShoot && mount.target instanceof Building b && b.damaged() && (heal.effectTimer += Time.delta) >= reload){
+        if(canShoot && heal.target instanceof Building b && b.damaged() && (heal.effectTimer += Time.delta) >= healEffectInterval){
             healEffect.at(b.x, b.y, 0f, healColor, b.block);
             heal.effectTimer = 0f;
         }
 
-        if(canShoot && mount.target instanceof Healthc u){
+        if(canShoot && heal.target instanceof Healthc u){
             float baseAmount = repairSpeed * heal.strength * Time.delta + fractionRepairSpeed * heal.strength * Time.delta * u.maxHealth() / 100f;
             u.heal((u instanceof Building b && b.wasRecentlyDamaged() ? recentDamageMultiplier : 1f) * baseAmount);
         }
     }
 
     @Override
-    public void draw(Unit unit, WeaponMount mount){
-        super.draw(unit, mount);
-
-        HealBeamMount heal = (HealBeamMount)mount;
-
-        if(unit.canShoot()){
-            float
-                weaponRotation = unit.rotation - 90,
-                wx = unit.x + Angles.trnsx(weaponRotation, x, y),
-                wy = unit.y + Angles.trnsy(weaponRotation, x, y),
-                z = Draw.z();
-            RepairTurret.drawBeam(wx, wy, unit.rotation + mount.rotation, shootY, unit.id, mount.target == null || controllable ? null : (Sized)mount.target, unit.team, heal.strength,
-            pulseStroke, pulseRadius, beamWidth + Mathf.absin(widthSinScl, widthSinMag), heal.lastEnd, heal.offset, laserColor, laserTopColor,
-            laser, laserEnd, laserTop, laserTopEnd);
-            Draw.z(z);
-        }
+    protected Teamc findTarget(Unit unit, float x, float y, float range, boolean air, boolean ground){
+        var out = targetUnits ? Units.closest(unit.team, x, y, range, u -> u != unit && u.damaged()) :  null;
+        if(out != null || !targetBuildings) return out;
+        return Units.findAllyTile(unit.team, x, y, range, Building::damaged);
     }
 
     @Override
-    public void init(){
-        super.init();
-        bullet.healPercent = fractionRepairSpeed;
+    protected boolean checkTarget(Unit unit, Teamc target, float x, float y, float range){
+        return !(target.within(unit, range + unit.hitSize/2f) && target.team() == unit.team && target instanceof Healthc u && u.damaged() && u.isValid());
     }
 
-    public static class HealBeamMount extends WeaponMount{
+    @Override
+    public void load(){
+        super.load();
+
+        laser = Core.atlas.find("laser-white");
+        laserEnd = Core.atlas.find("laser-white-end");
+        laserTop = Core.atlas.find("laser-top");
+        laserTopEnd = Core.atlas.find("laser-top-end");
+    }
+
+    public static class HealBeamMountR extends TargetWeaponMount{
         public Vec2 offset = new Vec2(), lastEnd = new Vec2();
         public float strength, effectTimer;
 
-        public HealBeamMount(Weapon weapon){
+        public HealBeamMountR(BaseWeapon weapon){
             super(weapon);
         }
     }
