--- core/src/mindustry/type/weapons/RepairBeamWeaponR.java
@@ -1,164 +1,0 @@
-package mindustry.type.weapons;
-
-import arc.*;
-import arc.graphics.*;
-import arc.graphics.g2d.*;
-import arc.math.*;
-import arc.math.geom.*;
-import arc.scene.ui.layout.*;
-import arc.util.*;
-import mindustry.*;
-import mindustry.content.*;
-import mindustry.core.*;
-import mindustry.entities.*;
-import mindustry.entities.units.*;
-import mindustry.gen.*;
-import mindustry.graphics.*;
-import mindustry.type.*;
-import mindustry.world.blocks.units.*;
-import mindustry.world.meta.*;
-
-import static mindustry.Vars.*;
-
-public class RepairBeamWeaponR extends TargetWeapon{
-    public boolean targetBuildings = false, targetUnits = true;
-
-    public float repairSpeed = 0.3f;
-    public float fractionRepairSpeed = 0f;
-    public float beamWidth = 1f;
-    public float pulseRadius = 6f;
-    public float pulseStroke = 2f;
-    public float widthSinMag = 0f, widthSinScl = 4f;
-    public float recentDamageMultiplier = 0.1f;
-
-    public TextureRegion laser, laserEnd, laserTop, laserTopEnd;
-
-    public Color laserColor = Color.valueOf("98ffa9"), laserTopColor = Color.white.cpy();
-    //only for blocks
-    public Color healColor = Pal.heal;
-    public Effect healEffect = Fx.healBlockFull;
-    public float healEffectInterval = 1f;
-
-    public RepairBeamWeaponR(String name){
-        super(name);
-    }
-
-    public RepairBeamWeaponR(){
-    }
-
-    {
-        autoTarget = true;
-        controllable = false;
-        rotate = true;
-        mountType = HealBeamMountR::new;
-        recoil = 0f;
-    }
-
-    @Override
-    public void addStats(UnitType u, Table w){
-        w.row();
-        w.add("[lightgray]" + Stat.repairSpeed.localized() + ": " + (mirror ? "2x " : "") + "[white]" + (int)(repairSpeed * 60) + " " + StatUnit.perSecond.localized());
-    }
-
-    @Override
-    public void drawWeapon(Unit unit, BaseWeaponMount mount, float wx, float wy, float weaponRotation){
-        super.drawWeapon(unit, mount, wx, wy, weaponRotation);
-
-        HealBeamMountR heal = (HealBeamMountR)mount;
-
-        if(unit.canShoot()){
-            float z = Draw.z();
-            RepairTurret.drawBeam(wx, wy, unit.rotation + mount.rotation, shootY, unit.id, heal.target == null || controllable ? null : (Sized)heal.target, unit.team, heal.strength,
-                pulseStroke, pulseRadius, beamWidth + Mathf.absin(widthSinScl, widthSinMag), heal.lastEnd, heal.offset, laserColor, laserTopColor,
-                laser, laserEnd, laserTop, laserTopEnd);
-            Draw.z(z);
-        }
-    }
-
-    @Override
-    public void update(Unit unit, BaseWeaponMount mount){
-        super.update(unit, mount);
-
-        float
-            weaponRotation = unit.rotation - 90,
-            wx = unit.x + Angles.trnsx(weaponRotation, x, y),
-            wy = unit.y + Angles.trnsy(weaponRotation, x, y);
-
-        HealBeamMountR heal = (HealBeamMountR)mount;
-        boolean canShoot = mount.shoot;
-
-        if(!autoTarget){
-            heal.target = null;
-            if(canShoot){
-                heal.lastEnd.set(heal.aimX, heal.aimY);
-
-                if(!rotate && !Angles.within(Angles.angle(wx, wy, heal.aimX, heal.aimY), unit.rotation, shootCone)){
-                    canShoot = false;
-                }
-            }
-
-            //limit range
-            heal.lastEnd.sub(wx, wy).limit(range()).add(wx, wy);
-
-            if(targetBuildings){
-                //snap to closest building
-                World.raycastEachWorld(wx, wy, heal.lastEnd.x, heal.lastEnd.y, (x, y) -> {
-                    var build = Vars.world.build(x, y);
-                    if(build != null && build.team == unit.team && build.damaged()){
-                        heal.target = build;
-                        heal.lastEnd.set(x * tilesize, y * tilesize);
-                        return true;
-                    }
-                    return false;
-                });
-            }
-            if(targetUnits){
-                //TODO does not support healing units manually yet
-            }
-        }
-
-        heal.strength = Mathf.lerpDelta(heal.strength, Mathf.num(autoTarget ? heal.target != null : canShoot), 0.2f);
-
-        //create heal effect periodically
-        if(canShoot && heal.target instanceof Building b && b.damaged() && (heal.effectTimer += Time.delta) >= healEffectInterval){
-            healEffect.at(b.x, b.y, 0f, healColor, b.block);
-            heal.effectTimer = 0f;
-        }
-
-        if(canShoot && heal.target instanceof Healthc u){
-            float baseAmount = repairSpeed * heal.strength * Time.delta + fractionRepairSpeed * heal.strength * Time.delta * u.maxHealth() / 100f;
-            u.heal((u instanceof Building b && b.wasRecentlyDamaged() ? recentDamageMultiplier : 1f) * baseAmount);
-        }
-    }
-
-    @Override
-    protected Teamc findTarget(Unit unit, float x, float y, float range, boolean air, boolean ground){
-        var out = targetUnits ? Units.closest(unit.team, x, y, range, u -> u != unit && u.damaged()) :  null;
-        if(out != null || !targetBuildings) return out;
-        return Units.findAllyTile(unit.team, x, y, range, Building::damaged);
-    }
-
-    @Override
-    protected boolean checkTarget(Unit unit, Teamc target, float x, float y, float range){
-        return !(target.within(unit, range + unit.hitSize/2f) && target.team() == unit.team && target instanceof Healthc u && u.damaged() && u.isValid());
-    }
-
-    @Override
-    public void load(){
-        super.load();
-
-        laser = Core.atlas.find("laser-white");
-        laserEnd = Core.atlas.find("laser-white-end");
-        laserTop = Core.atlas.find("laser-top");
-        laserTopEnd = Core.atlas.find("laser-top-end");
-    }
-
-    public static class HealBeamMountR extends TargetWeaponMount{
-        public Vec2 offset = new Vec2(), lastEnd = new Vec2();
-        public float strength, effectTimer;
-
-        public HealBeamMountR(BaseWeapon weapon){
-            super(weapon);
-        }
-    }
-}
